# doc/openapi/openapi.yml
openapi: '3.0.3'
info:
    title: ids
    version: '0.1.0'
servers:
  - url: api/v1
paths:
...
    /teachers:
        get:
            operationId: get-teachers
            parameters:
                - $ref: '#/components/parameters/TeachersStateQuery'
                - $ref: '#/components/parameters/PagenationLimitQuery'
                - $ref: '#/components/parameters/PagenationOffsetQuery'
            responses:
                '200':
                    $ref: '#/components/responses/OkGetTeachersList'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
            tags:
              - teachers
        
        post:
            operationId: post-teachers
            requestBody:
                $ref: '#/components/requestBodies/PostTeachers'
            responses:
                '201':
                    $ref: '#/components/responses/OkPostTeachers'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
            tags:
              - teachers
    /teachers/{teacher_id}:
        get:
            operationId: get-teacher
            parameters:
                - $ref: '#/components/parameters/TeacherId'
            responses:
                '200':
                    $ref: '#/components/responses/OkGetTeacher'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
            tags:
              - teachers
        
        put:
            operationId: put-Teacher
            parameters:
                - $ref: '#/components/parameters/TeacherId'
            requestBody:
                $ref: '#/components/requestBodies/PutTeacher'
            responses:
                '200':
                    $ref: '#/components/responses/OkPutTeacher'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
            tags:
              - teachers

        patch:
            operationId: patch-Teacher
            parameters:
                - $ref: '#/components/parameters/TeacherId'
            requestBody:
                $ref: '#/components/requestBodies/PatchTeacher'
            responses:
                '200':
                    $ref: '#/components/responses/OkPatchTeacher'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
            tags:
              - teachers
              
        delete:
            operationId: delete-Teacher
            parameters:
                - $ref: '#/components/parameters/TeacherId'
            responses:
                '204':
                    description: The teacher has been successfully deleted
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
            tags:
              - teachers
components:
...
    schemas:
        PagenationLimit:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
        PagenationOffset:
            type: integer
        UserId:
            type: string
            format: uuid
...
        TeacherId:
            type: string
            format: uuid
        TeachersStatus:
            type: string
            enum: [active, inactive]
...
        Teacher:
            properties:
                id:
                    $ref: '#/components/schemas/TeacherId'
                userId:
                    $ref: '#/components/schemas/UserId'
                lastName:
                    type: string
                firstName:
                    type: string
                lastNameKana:
                    type: string
                firstNameKana:
                    type: string
                phone:
                    type: string
                mobilePhone:
                    type: string
                email:
                    type: string
                postCode1:
                    type: string
                postCode2:
                    type: string
                prefecture:
                    type: string
                city:
                    type: string
                streetAddress:
                    type: string
                building:
                    type: string
                prefecturesKana:
                    type: string
                cityKana:
                    type: string
                buildingKana:
                    type: string
                hireDate:
                    type: string
                    format: date
                leaveDate:
                    type: string
                    format: date
                status:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
            required:
              - id
              - userId
              - lastName
              - firstName
              - postCode1
              - postCode2
              - prefectures
              - city
              - streetAddress
              - hireDate
              - status
              - createdAt
              - updatedAt
        TeacherList:
            type: array
            items:
                $ref: '#/components/schemas/Teacher'
...
        RequestPostTeachers:
            type: object
            properties:
                lastName:
                    type: string
                firstName:
                    type: string
                lastNameKana:
                    type: string
                firstNameKana:
                    type: string
                phone:
                    type: string
                mobilePhone:
                    type: string
                email:
                    type: string
                postCode1:
                    type: string
                postCode2:
                    type: string
                prefecture:
                    type: string
                city:
                    type: string
                streetAddress:
                    type: string
                building:
                    type: string
                prefecturesKana:
                    type: string
                cityKana:
                    type: string
                buildingKana:
                    type: string
                hireDate:
                    type: string
                    format: date
            required:
                - lastName
                - firstName
                - postCode1
                - postCode2
                - prefectures
                - city
                - streetAddress
                - hireDate
        RequestPutTeacher:
            type: object
            properties:
                lastName:
                    type: string
                firstName:
                    type: string
                lastNameKana:
                    type: string
                firstNameKana:
                    type: string
                phone:
                    type: string
                mobilePhone:
                    type: string
                email:
                    type: string
                postCode1:
                    type: string
                postCode2:
                    type: string
                prefecture:
                    type: string
                city:
                    type: string
                streetAddress:
                    type: string
                building:
                    type: string
                prefecturesKana:
                    type: string
                cityKana:
                    type: string
                buildingKana:
                    type: string
                hireDate:
                    type: string
                    format: date
            required:
                - lastName
                - firstName
                - postCode1
                - postCode2
                - prefectures
                - city
                - streetAddress
                - hireDate
        RequestPatchTeacher:
            type: object
            properties:
                leaveDate:
                    type: string
                    format: date
                status:
                    type: string
            required:
                - status

    #-------------------------------
    # Reusable operation parameters
    #-------------------------------
    parameters:
        PagenationLimitQuery:
            name: limit
            in: query
            required: false
            schema:
              $ref: '#/components/schemas/PagenationLimit'
        PagenationOffsetQuery:
            name: offset
            in: query
            required: false
            schema:
              $ref: '#/components/schemas/PagenationOffset'
...
        TeachersStateQuery:
            name: status
            in: query
            required: false
            schema:
                $ref: '#/components/schemas/TeachersStatus'
        TeacherId:
            name: teacher_id
            in: path
            required: true
            schema:
                $ref: '#/components/schemas/TeacherId'
    
    #-------------------------------
    # Reusable request bodies
    #-------------------------------
    requestBodies:
...
        PostTeachers:
            description: post teachers
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/RequestPostTeachers'
        PutTeacher:
            description: put teacher
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/RequestPutTeacher'
        PatchTeacher:
            description: patch teacher
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/RequestPatchTeacher'

    #-------------------------------
    # Reusable responses bodies
    #-------------------------------
    responses:
...
        OkGetTeachersList:
            description: OK
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            items:
                                $ref: '#/components/schemas/TeacherList'
                            offset:
                                $ref: '#/components/schemas/PagenationOffset'
                            limit:
                                $ref: '#/components/schemas/PagenationLimit'
                            total:
                                type: integer
        OkPostTeachers:
            description: OK
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/TeacherId'
        OkGetTeacher:
            description: OK
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Teacher'
        OkPutTeacher:
            description: OK
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/TeacherId'
        OkPatchTeacher:
            description: OK
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/TeacherId'
        NotFound:
            description: not found
        Unauthorized:
            description: unauthorized
    securitySchemes:
        token:
            type: http
            scheme: bearer
            bearerFormat: JWT
security:
- token: []
...